requirements.txt
django~=4.0.4
django_rest_framework
djangorestframework-simplejwt
django-cors-headers
drf_yasg
django-rest-swagger
psycopg2
djangorestframework~=3.13.1
django-filter==21.1
PyJWT~=2.4.0
djoser